Bugfixes, Updates to 1.5.5e

+++ UPLOAD/zc_install/includes/classes/class.zcDatabaseInstaller.php
@@ -2,9 +2,9 @@
 /**
  * file contains zcDatabaseInstaller Class
  * @package Installer
- * @copyright Copyright 2003-2016 Zen Cart Development Team
+ * @copyright Copyright 2003-2017 Zen Cart Development Team
  * @license http://www.zen-cart-pro.at/license/2_0.txt GNU Public License V2.0
- * @version $Id: class.zcDatabaseInstaller.php  2016-03-27 18:59:53Z webchills $
+ * @version $Id: class.zcDatabaseInstaller.php 2017-03-05 15:59:53Z webchills $
  *
  */
 /**
@@ -37,7 +37,7 @@ class zcDatabaseInstaller
     $this->dieOnErrors = isset($options['dieOnErrors']) ? (bool)$options['dieOnErrors'] : FALSE;
     $this->errors = array();
     $this->basicParseStrings = array(
-    'DROP TABLE IF EXISTS ',
+    'DROP TABLE ',
     'CREATE TABLE ',
     'REPLACE INTO ',
     'INSERT INTO ',
@@ -103,6 +103,7 @@ class zcDatabaseInstaller
       if ( substr($this->line,-1) ==  ';')
       {
         if (substr($this->newLine,-1)==' ') $this->newLine = substr($this->newLine,0,(strlen($this->newLine)-1));
+        if (substr($this->newLine,-1)==')') $this->newLine = substr($this->newLine,0,(strlen($this->newLine)-1)) . ' )';
         $this->keepTogetherCount++;
         if ($this->keepTogetherCount == $this->keepTogetherLines)
         {
@@ -163,9 +164,23 @@ class zcDatabaseInstaller
       error_log("MySQL error " . $this->db->error_number . " encountered during zc_install:\n" . $this->db->error_text . "\n" . $this->line . "\n---------------\n\n");
     }
   }
-  public function parserDropTableIfExists ()
+  public function parserDropTable()
   {
-    $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+    $table = (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) == 'IF EXISTS') ? $this->lineSplit[4] : $this->lineSplit[2];
+
+    if (!$this->tableExists($table) && (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS'))
+    {
+      $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS')
+      {
+        $this->line = 'DROP TABLE ' . $this->dbPrefix . substr($this->line, 11);
+      } else {
+        $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+      }
+    }
   }
   public function parserCreateTable()
   {
@@ -325,6 +340,24 @@ class zcDatabaseInstaller
             break;
           default:
             // Do nothing
+      }
+    }
+  }
+  public function parserRenameTable()
+  {
+    if (!$this->tableExists($this->lineSplit[2])) 
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if ($this->tableExists($this->lineSplit[4]))
+      {
+        if (!isset($result)) $result = sprintf(REASON_TABLE_ALREADY_EXISTS, $table);
+        $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+        $this->ignoreLine = true;
+      } else {
+        $this->line = 'RENAME TABLE ' . $this->dbPrefix . $this->lineSplit[2] . ' TO ' . $this->dbPrefix . substr($this->line, (13 + strlen($this->lineSplit[2]) + 4));
       }
     }
   }
